fix adding power consumption characteristic
HomeKit bombs out if a characteristic doesn't have an `iid` set, and
one isn't assigned if you create a characteristic after a service has
been added to an accessory.  Fix this by manually creating the switch
accessory, adding the characteristic to the service and then adding
the service to the accessory.